+++ src/routes/venue/[id]/+page.ts
@@ -6,26 +6,48 @@ export const load: PageLoad = async ({ parent, params }) => {
 	const venueid = params.id;
 
 	// Get the matching venue.
-	const { data: venue } = await supabase.from('venues').select().eq('id', venueid).single();
+	const { data: venue, error: venueError } = await supabase
+		.from('venues')
+		.select()
+		.eq('id', venueid)
+		.single();
+	if (venueError) console.log(venueError);
+
 	// Get the matching venu's currency.
-	const { data: currency } = venue
+	const { data: currency, error: currencyError } = venue
 		? await supabase.from('currencies').select().eq('id', venue.currency).single()
 		: { data: null };
+	if (currencyError) console.log(currencyError);
+
 	// Get the matching venue's roles.
-	const { data: roles } = await supabase.from('roles').select().eq('venueid', venueid);
+	const { data: roles, error: rolesError } = await supabase
+		.from('roles')
+		.select()
+		.eq('venueid', venueid);
+	if (rolesError) console.log(rolesError);
+
 	// Get any commitments the user has made to the venue.
-	const { data: commitments, error } = user
+	const { data: commitments, error: commitmentsError } = user
 		? await supabase
 				.from('volunteers')
 				.select('*, roles (venueid)')
 				.eq('roles.venueid', venueid)
 				.eq('scholarid', user.id)
 		: { data: null };
+	if (commitmentsError) console.log(commitmentsError);
+
+	// See how many tokens the venue posseses.
+	const { data: tokens, error: tokensError } = await supabase
+		.from('tokens')
+		.select()
+		.eq('venue', venueid);
+	if (tokensError) console.log(tokensError);
 
 	return {
 		venue,
 		currency,
 		roles,
-		commitments
+		commitments,
+		tokens: tokens?.length ?? null
 	};
 };